requirements.txt
requests
win10toast-withsound;platform_system=="Windows"
notifypy;platform_system=="Linux" or platform_system=="Darwin"
colorama
playsound;platform_system=="Windows"
jsonc-parser